moved App to index + cosmetic changes
